--- frontend/src/components/PluginSearch/index.ts
@@ -1,1 +1,0 @@
-export * from './PluginSearch';
