+++ frontend/src/components/SearchPage/PluginComplexFilter.module.scss
