+++ frontend/src/components/SearchPage/PluginFilterByForm.tsx
