+++ frontend/src/components/SearchPage/PluginSearchBar.tsx
