+++ frontend/src/components/SearchPage/PluginSearchControls.tsx
