+++ frontend/src/components/SearchPage/PluginSearchResult.tsx
