+++ frontend/src/components/SearchPage/PluginSearchResultList.tsx
