+++ frontend/src/components/SearchPage/index.ts
@@ -1,0 +1,1 @@
+export * from './SearchPage';
