+++ frontend/src/components/SearchPage/useFilterLabels.ts
